refactor(desktop): satisfy file size policy
Move the event lookup helpers into focused modules so the existing oversized API files do not grow.
